Apply Player name formatting to DMs in dscord embeds, move the helper function to a utility location.
